revert offline node bag and buffer completion handling logic
this was essentially how it was before we merged #20.
in that, even if the bag is finished, if the registration buffers are
not empty, we still try to pop from the bag although its empty.
obviously leads to seg faults. until we can figure out a better way to
handle this, this will have to be