in case initialization is off, the second lidar frame timestamp was never updated.
this can be bad if the platform actually starts from motion, and
nevertheless is a mistake. fixed now.

the first frame which is used as global frame is also now logged in the
trajectory dump as identity

plus run_name has a better default value by taking the name of the data
itself.