remove the ament cmake find outside of a ros build, as that causes issues
in case a ros workspace is overlaid, find ament_camke will always
trigger. and in case the ros workspace is ill-configured (like mine),
even if one doesnt need it, any error thrown during the ament find can
propagate up and still terminate the build. not ideal. this is the
nuclear option to deal with this problem. ros builds are opt-out, i.e., enabled
by default. pyproject handles the necessary flags for python