move preproc to its own file and potentially fix the non intialization behaviour
for the map. now even the first scan gets preproc'd, which might be a
bit wasteful, but its a one time thing and keeps things simple. if
init is off, map gets updated with the map update frame so that icp
next time doesn't fall through. behaviour should remain the same as
before if init is on